# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Return API
  version: 1.1.0
  description: |
    **Return** — API de gestion de prêts d'objets personnels et d'argent.

    Cette API permet de :
    - Gérer des prêts d'objets (photo, retour physique) et d'argent (montant, remboursement)
    - Suivre le cycle de vie complet d'un prêt (création, confirmation, rappels, retour)
    - Planifier et envoyer des rappels automatiques diplomatiques (J-3, J, J+7, J+14, J+21)
    - Consulter l'historique et les statistiques de prêts

    **Architecture** : Monolithe Modulaire (NestJS 11.x + PostgreSQL 17+)
    **Authentification** : JWT (Bearer Token) avec révocation immédiate via Redis
    **Standards** : RFC 7807 pour toutes les erreurs, i18n FR/EN
    **Validation** : class-validator + class-transformer

  contact:
    name: Return Support
    email: support@return.app
    url: https://return.app/support
  license:
    name: Proprietary
    url: https://return.app/license
servers:
  - url: https://api.return.app/v1
    description: Production (Fly.io — Europe West)
  - url: http://localhost:3000/v1
    description: Développement local
...
security:
  - BearerAuth: []
paths:
...
  /items:
    get:
      tags:
        - Items
      summary: Lister tous les objets de l'utilisateur
      description: |
        Récupère tous les objets créés par l'utilisateur, qu'ils soient actuellement
        prêtés ou non.
      operationId: listItems
      parameters:
        - name: category
          in: query
          description: Filtrer par catégorie
          schema:
            $ref: '#/components/schemas/ItemCategory'
          example: TOOLS
        - name: available
          in: query
          description: Filtrer les objets disponibles (non prêtés actuellement)
          schema:
            type: boolean
          example: true
        - name: page
          in: query
          description: Numéro de page (commence à 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Nombre de résultats par page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Liste des objets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Items
      summary: Créer un objet
      description: |
        Créer un objet par saisie manuelle (nom, catégorie, valeur estimée).
        Une photo peut être ajoutée séparément via `/items/{itemId}/photos`.
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemDto'
      responses:
        '201':
          description: Objet créé avec succès
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: https://api.return.app/v1/items/9a1b2c3d-4e5f-6a7b-8c9d-0e1f2a3b4c5d
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /items/{itemId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Items
      summary: Récupérer les détails d'un objet
      description: Retourne toutes les informations d'un objet, y compris ses photos et son historique de prêts.
      operationId: getItemById
      responses:
        '200':
          description: Détails de l'objet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Items
      summary: Mettre à jour un objet
      description: Permet de modifier le nom, la description, la catégorie ou la valeur estimée d'un objet existant.
      operationId: updateItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 100
                description:
                  type: string
                  nullable: true
                  maxLength: 500
                category:
                  $ref: '#/components/schemas/ItemCategory'
                estimatedValue:
                  type: number
                  format: float
                  nullable: true
                  minimum: 0
      responses:
        '200':
          description: Objet mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Items
      summary: Supprimer un objet
      description: |
        Supprime un objet s'il n'est associé à aucun prêt actif.
        Si l'objet est prêté, la suppression est refusée (409 Conflict).
      operationId: deleteItem
      responses:
        '204':
          description: Objet supprimé avec succès
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: L'objet est actuellement prêté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://api.return.app/errors/item-currently-loaned
                title: Item Currently Loaned
                status: 409
                detail: Cannot delete an item that is currently on loan.
                instance: /v1/items/item-123
                timestamp: 2026-02-08T18:30:00Z
                requestId: req-0d1e2f3a
  /items/{itemId}/photos:
    parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags:
        - Items
      summary: Ajouter une photo à un objet existant
      description: |
        Upload d'une photo supplémentaire (max 5 photos par objet).
      operationId: addPhotoToItem
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: Image au format JPEG ou PNG (max 5MB)
      responses:
        '201':
          description: Photo ajoutée avec succès
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: https://api.return.app/v1/items/9a1b2c3d-4e5f-6a7b-8c9d-0e1f2a3b4c5d/photos/1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Format invalide ou limite de photos atteinte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://api.return.app/errors/max-photos-exceeded
                title: Maximum Photos Exceeded
                status: 400
                detail: An item can have a maximum of 5 photos.
                instance: /v1/items/item-123/photos
                timestamp: 2026-02-08T18:40:00Z
                requestId: req-1e2f3a4b
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /items/{itemId}/photos/{photoId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      tags:
        - Items
      summary: Supprimer une photo d'un objet
      description: |
        Supprime une photo associée à un objet. Le fichier est supprimé du storage
        et l'enregistrement est retiré de la base de données.
      operationId: deletePhotoFromItem
      responses:
        '204':
          description: Photo supprimée avec succès
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # BORROWERS ENDPOINTS
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtenu via `/auth/login` ou `/auth/register`.
        Format : `Authorization: Bearer <access_token>`

        Le token expire après 15 minutes. Utiliser `/auth/refresh` pour le renouveler.
  schemas:
...
    Item:
      type: object
      description: Bien physique ou somme d'argent faisant l'objet d'un Prêt
      required:
        - id
        - name
        - category
      properties:
        id:
          type: string
          format: uuid
          example: 9a1b2c3d-4e5f-6a7b-8c9d-0e1f2a3b4c5d
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: Perceuse Bosch PSB 1800
          description: Nom de l'objet (saisie manuelle)
        description:
          type: string
          nullable: true
          maxLength: 500
          example: Perceuse-visseuse sans fil 18V avec 2 batteries Li-Ion
        category:
          $ref: '#/components/schemas/ItemCategory'
        estimatedValue:
          type: number
          format: float
          nullable: true
          minimum: 0
          example: 129.99
          description: |
            Valeur estimée de l'objet en euros (optionnelle).
            **Obligatoire** si category = MONEY (constitue alors le montant prêté).
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
          description: Photos de l'objet (max 5)
          maxItems: 5
        createdAt:
          type: string
          format: date-time
          example: '2026-02-08T14:00:00Z'
    ItemCategory:
      type: string
      enum:
        - TOOLS
        - BOOKS
        - ELECTRONICS
        - SPORTS
        - KITCHEN
        - GARDEN
        - MONEY
        - OTHER
      description: |
        Catégorie de l'objet sélectionnée manuellement par le prêteur.
        **MONEY** active le workflow de prêt d'argent (estimatedValue = montant prêté, obligatoire).
        Les autres catégories concernent des prêts d'objets physiques (workflow standard avec photo).
      example: TOOLS
    Photo:
      type: object
      description: Photo d'un objet stockée sur Cloudflare R2
      required:
        - id
        - url
        - uploadedAt
      properties:
        id:
          type: string
          format: uuid
          example: 1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d
        url:
          type: string
          format: uri
          example: https://photos.return.app/loans/loan-123/photo-1.jpg
          description: URL publique de la photo (signée, expire après 7 jours)
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
          example: https://photos.return.app/loans/loan-123/photo-1-thumb.jpg
          description: URL de la vignette optimisée (150x150px)
        uploadedAt:
          type: string
          format: date-time
          example: '2026-02-08T14:05:00Z'
...
    PaginationMetadata:
      type: object
      description: Métadonnées de pagination pour les listes paginées
      required:
        - currentPage
        - totalPages
        - totalItems
        - itemsPerPage
      properties:
        currentPage:
          type: integer
          example: 1
          description: Page actuelle (commence à 1)
        totalPages:
          type: integer
          example: 5
          description: Nombre total de pages
        totalItems:
          type: integer
          example: 87
          description: Nombre total d'éléments
        itemsPerPage:
          type: integer
          example: 20
          description: Nombre d'éléments par page
        hasNextPage:
          type: boolean
          example: true
          description: Y a-t-il une page suivante ?
        hasPreviousPage:
          type: boolean
          example: false
          description: Y a-t-il une page précédente ?
...
    CreateItemDto:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: Perceuse Bosch PSB 1800
        description:
          type: string
          nullable: true
          maxLength: 500
          example: Perceuse-visseuse sans fil 18V
        category:
          $ref: '#/components/schemas/ItemCategory'
        estimatedValue:
          type: number
          format: float
          nullable: true
          minimum: 0
          example: 129.99
...
    ProblemDetails:
      type: object
      description: |
        Format standardisé RFC 7807 pour toutes les erreurs API.
        Documentation : https://tools.ietf.org/html/rfc7807
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - requestId
      properties:
        type:
          type: string
          format: uri
          example: https://api.return.app/errors/loan-not-found
          description: URI identifiant le type d'erreur
        title:
          type: string
          example: Loan Not Found
          description: Titre court lisible par un humain
        status:
          type: integer
          example: 404
          description: Code HTTP de l'erreur
        detail:
          type: string
          example: "The loan with ID 'loan-123' does not exist or you don't have access to it."
          description: Explication détaillée de l'erreur
        instance:
          type: string
          example: /v1/loans/loan-123
          description: URI de la requête ayant causé l'erreur
        timestamp:
          type: string
          format: date-time
          example: '2026-02-08T14:32:00Z'
          description: Date/heure de l'erreur
        requestId:
          type: string
          example: req-7f3c9a2b-4d1e-4a8f-9c7b-1e3f5a6b8c9d
          description: ID unique de la requête (pour traçabilité)
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Liste des erreurs de validation (optionnel)
    ErrorDetail:
      type: object
      description: Détail d'une erreur de validation sur un champ spécifique
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          example: returnDate
          description: Nom du champ en erreur
        code:
          type: string
          example: INVALID_DATE
          description: Code d'erreur (utilisable pour i18n)
        message:
          type: string
          example: Return date must be in the future
          description: Message d'erreur lisible

  # ===========================================================================
  # REUSABLE RESPONSES
  # ===========================================================================
  responses:
    Unauthorized:
      description: Non authentifié (token manquant ou invalide)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            missingToken:
              value:
                type: https://api.return.app/errors/unauthorized
                title: Unauthorized
                status: 401
                detail: No authorization token was found in the request headers.
                instance: /v1/loans
                timestamp: 2026-02-08T14:32:00Z
                requestId: req-9b8c7d6e
            invalidToken:
              value:
                type: https://api.return.app/errors/invalid-token
                title: Invalid Token
                status: 401
                detail: The provided token is invalid or has expired.
                instance: /v1/loans
                timestamp: 2026-02-08T14:35:00Z
                requestId: req-1a2b3c4d
    Forbidden:
      description: Non autorisé (permissions insuffisantes)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            notOwner:
              value:
                type: https://api.return.app/errors/forbidden
                title: Forbidden
                status: 403
                detail: You can only access or modify loans you created.
                instance: /v1/loans/loan-123
                timestamp: 2026-02-08T14:40:00Z
                requestId: req-5e6f7a8b
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            loanNotFound:
              value:
                type: https://api.return.app/errors/loan-not-found
                title: Loan Not Found
                status: 404
                detail: "The loan with ID 'loan-999' does not exist."
                instance: /v1/loans/loan-999
                timestamp: 2026-02-08T14:45:00Z
                requestId: req-9c8d7e6f
    ValidationError:
      description: Erreur de validation des données d'entrée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            multipleErrors:
              value:
                type: https://api.return.app/errors/validation-failed
                title: Validation Failed
                status: 400
                detail: The request contains invalid data.
                instance: /v1/auth/register
                timestamp: 2026-02-08T14:50:00Z
                requestId: req-1f2e3d4c
                errors:
                  - field: email
                    code: INVALID_EMAIL
                    message: Email format is invalid
                  - field: password
                    code: TOO_SHORT
                    message: Password must be at least 8 characters
...